% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnonymizationApp.R
\docType{data}
\name{AnonymizationApp}
\alias{AnonymizationApp}
\alias{ui}
\title{AnonymizationApp}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
ui
}
\value{
An anonymized file which is a copy of the origin file but without the personal information and another filename
}
\description{
Takes an .fcs or .lmd file as an input, anonymizes all the personal data in that file and save it as a new .fcs or .lmd file within a Shiny App
}
\keyword{datasets}
