% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnonymizationApp.R
\name{AnonymizationApp}
\alias{AnonymizationApp}
\title{AnonymizationApp}
\arguments{
\item{Btn_GetFile}{Input of the filepath of the file that is about to be made anonymous}

\item{newFileName}{the name of the new file which is a copy of the origin file}

\item{goButton}{Submit Button to submit the new filename}

\item{Anonymized}{action Button to anonymize the new file}
}
\value{
An anonymized file which is a copy of the origin file but without the personal information and another filename
}
\description{
Takes an .fcs or .lmd file as an input, anonymizes all the personal data in that file and save it as a new .fcs or .lmd file within a Shiny App
}
